+++ ngx-numeric-range-form-field/projects/ngx-numeric-range-form-field/src/lib/form/numeric-range.validator.ts
@@ -1,13 +1,15 @@
 import { AbstractControl, ValidationErrors, ValidatorFn } from '@angular/forms';
 
-export const numericRangeValues: ValidatorFn = (
+export const numericRangeValues = (
+	minimumControlName = 'minimum',
+	maximumControlName = 'maximum') => (
 	group: AbstractControl
 ): ValidationErrors | null => {
-	const max = group.get('maximum').value
-		? Number(group.get('maximum').value)
+	const max = group.get(maximumControlName).value
+		? Number(group.get(maximumControlName).value)
 		: null;
-	const min = group.get('minimum').value
-		? Number(group.get('minimum').value)
+	const min = group.get(minimumControlName).value
+		? Number(group.get(minimumControlName).value)
 		: null;
 
 	if (max !== null && min !== null) {